requirements.txt
tinyec==0.3.1
pycryptodome==3.10.1
